DS-296 - Make geo represention for 'HR vestigingen' in CSV the same as for 'adressen'
lat/lon are printed as strings with reral comma separator, and th RD values are also shown
